validate API key for config APIs
