+++ lib/utils.ts
@@ -1,0 +1,16 @@
+import { NextApiRequest } from 'next';
+import env from '@lib/env';
+
+export const validateApiKey = (token) => {
+  return env.apiKeys.includes(token);
+};
+
+export const extractAuthToken = (req: NextApiRequest) => {
+  const authHeader = req.headers['authorization'];
+  const parts = (authHeader || '').split(' ');
+  if (parts.length > 1) {
+    return parts[1];
+  }
+
+  return null;
+};
